Port to Create 6.x and Flywheel 1.x
As far as my testing goes, this seems to work. I haven't tested this
exhaustively as far as the Create compat but I figured I'd give this
source away so others can test it more thoroughly than I have, and
give the mod author a head start since it is open source and this mod
is so crucial to many players' experiences (including mine).